changed project name from testplug to 'sbmlsheets'
