+++ bundles/org.insilico.sbmlsheets/META-INF/MANIFEST.MF
@@ -1,0 +1,22 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: SBMLsheets
+Bundle-SymbolicName: org.insilico.sbmlsheets
+Bundle-Version: 1.0.0.qualifier
+Require-Bundle: org.eclipse.core.runtime,
+ org.eclipse.fx.ui.services;bundle-version="3.0.0",
+ org.eclipse.fx.code.editor;bundle-version="3.0.0",
+ org.sbml.jsbml.core;bundle-version="1.3.1",
+ org.eclipse.fx.code.editor.fx.e4;bundle-version="3.0.0",
+ org.eclipse.e4.core.contexts;bundle-version="1.5.0",
+ org.eclipse.e4.core.di;bundle-version="1.6.1",
+ org.eclipse.e4.ui.model.workbench;bundle-version="1.2.0"
+Bundle-RequiredExecutionEnvironment: JavaSE-1.8
+Bundle-ActivationPolicy: lazy
+Service-Component: OSGI-INF/services/org.insilico.sbmlsheets.services.provider.SBMLEditorProvider.xml,
+ OSGI-INF/services/org.insilico.sbmlsheets.services.provider.SBMLIconProvider.xml,
+ OSGI-INF/services/org.insilico.sbmlsheets.services.provider.SBMLDocumentLoader.xml
+Export-Package: org.insilico.sbmlsheets,
+ org.insilico.sbmlsheets.editor,
+ org.insilico.sbmlsheets.services,
+ org.insilico.sbmlsheets.services.provider
