+++ bundles/org.insilico.sbmlsheets/OSGI-INF/services/org.insilico.sbmlsheets.services.provider.SBMLDocumentLoader.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" name="org.insilico.sbmlsheets.services.provider.SBMLDocumentLoader">
+   <property name="service.context.key" value="org.sbml.jsbml.SBMLDocument"/>
+   <service>
+      <provide interface="org.eclipse.e4.core.contexts.IContextFunction"/>
+   </service>
+   <implementation class="org.insilico.sbmlsheets.services.provider.SBMLDocumentLoader"/>
+</scr:component>
