+++ bundles/org.insilico.sbmlsheets/OSGI-INF/services/org.insilico.sbmlsheets.services.provider.SBMLIconProvider.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" name="org.insilico.sbmlsheets.services.provider.SBMLIconProvider">
+   <service>
+      <provide interface="org.eclipse.fx.code.editor.services.FileIconProvider"/>
+   </service>
+   <implementation class="org.insilico.sbmlsheets.services.provider.SBMLIconProvider"/>
+</scr:component>
