+++ bundles/org.insilico.sbmlsheets/res/sbml-logo-24.png
