+++ bundles/org.insilico.sbmlsheets/res/sbml-logo-48.png
