+++ bundles/org.insilico.sbmlsheets/src/org/insilico/sbmlsheets/Constants.java
@@ -1,0 +1,5 @@
+package org.insilico.sbmlsheets;
+
+public interface Constants {
+
+}
