+++ bundles/org.insilico.sbmlsheets/src/org/insilico/sbmlsheets/editor/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * 
+ */
+/**
+ * @author robert
+ *
+ */
+package org.insilico.sbmlsheets.editor;
