+++ bundles/org.insilico.sbmlsheets/src/org/insilico/sbmlsheets/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * <h1>Testplug</h1>
+ * 
+ * This bundle provides basic support for the SBML core specifications.
+ * 
+ * @see org.sbml.jsbml
+ */
+package org.insilico.sbmlsheets;
