+++ bundles/org.insilico.sbmlsheets/src/org/insilico/sbmlsheets/services/package-info.java
@@ -1,0 +1,10 @@
+/**
+ * This package contains service definitions (interfaces) which are related to SBML and/or the JSBML
+ * library.
+ * 
+ * @author robert
+ * @see org.insilico.sbmlsheets
+ * @see org.sbml.jsbml
+ *
+ */
+package org.insilico.sbmlsheets.services;
