+++ bundles/org.insilico.sbmlsheets/src/org/insilico/sbmlsheets/services/provider/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * This package contains OSGi service components which provides other services.
+ * 
+ * @author robert
+ * @see org.insilico.sbmlsheets
+ */
+package org.insilico.sbmlsheets.services.provider;
