--- bundles/org.insilico.testplug/META-INF/MANIFEST.MF
@@ -1,22 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ManifestVersion: 2
-Bundle-Name: Testplug
-Bundle-SymbolicName: org.insilico.testplug
-Bundle-Version: 1.0.0.qualifier
-Require-Bundle: org.eclipse.core.runtime,
- org.eclipse.fx.ui.services;bundle-version="3.0.0",
- org.eclipse.fx.code.editor;bundle-version="3.0.0",
- org.sbml.jsbml.core;bundle-version="1.3.1",
- org.eclipse.fx.code.editor.fx.e4;bundle-version="3.0.0",
- org.eclipse.e4.core.contexts;bundle-version="1.5.0",
- org.eclipse.e4.core.di;bundle-version="1.6.1",
- org.eclipse.e4.ui.model.workbench;bundle-version="1.2.0"
-Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Bundle-ActivationPolicy: lazy
-Service-Component: OSGI-INF/services/org.insilico.testplug.services.provider.SBMLEditorProvider.xml,
- OSGI-INF/services/org.insilico.testplug.services.provider.SBMLIconProvider.xml,
- OSGI-INF/services/org.insilico.testplug.services.provider.SBMLDocumentLoader.xml
-Export-Package: org.insilico.testplug,
- org.insilico.testplug.editor,
- org.insilico.testplug.services,
- org.insilico.testplug.services.provider
