--- bundles/org.insilico.testplug/OSGI-INF/services/org.insilico.testplug.services.provider.SBMLEditorProvider.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" name="org.insilico.testplug.services.provider.SBMLEditorProvider">
-   <service>
-      <provide interface="org.eclipse.fx.code.editor.fx.e4.EditorClassURLProvider"/>
-   </service>
-   <implementation class="org.insilico.testplug.services.provider.SBMLEditorProvider"/>
-</scr:component>
