--- bundles/org.insilico.testplug/src/org/insilico/testplug/Constants.java
@@ -1,5 +1,0 @@
-package org.insilico.testplug;
-
-public interface Constants {
-
-}
