--- bundles/org.insilico.testplug/src/org/insilico/testplug/SBMLUtils.java
@@ -1,19 +1,0 @@
-package org.insilico.testplug;
-
-/**
- * Util class for JSBML.
- * 
- * @author robert
- *
- */
-public final class SBMLUtils {
-    /**
-     * Checks if the file at the given uri is a valid SBML file.
-     * 
-     * @param uri
-     * @return <code>true</code> if the file is a sbml file.
-     */
-    public static final boolean isSBMLFile(String uri) {
-        return uri.endsWith("sbml");
-    }
-}
