--- bundles/org.insilico.testplug/src/org/insilico/testplug/editor/package-info.java
@@ -1,8 +1,0 @@
-/**
- * 
- */
-/**
- * @author robert
- *
- */
-package org.insilico.testplug.editor;
