--- bundles/org.insilico.testplug/src/org/insilico/testplug/package-info.java
@@ -1,8 +1,0 @@
-/**
- * <h1>Testplug</h1>
- * 
- * This bundle provides basic support for the SBML core specifications.
- * 
- * @see org.sbml.jsbml
- */
-package org.insilico.testplug;
