--- bundles/org.insilico.testplug/src/org/insilico/testplug/services/SBMLAdapterBuilderProvider.java
@@ -1,17 +1,0 @@
-package org.insilico.testplug.services;
-
-import org.sbml.jsbml.SBMLDocument;
-
-/**
- * A {@link SBMLAdapterBuilderProvider} provides a builder which can be used to create a product of
- * type P. The builder is used to form a adapter.
- * 
- * @author robert
- *
- * @param <P>
- */
-public interface SBMLAdapterBuilderProvider<P> {
-    public boolean canBuildProductFrom(SBMLDocument doc);
-
-    public P getBuilder();
-}
