--- bundles/org.insilico.testplug/src/org/insilico/testplug/services/package-info.java
@@ -1,10 +1,0 @@
-/**
- * This package contains service definitions (interfaces) which are related to SBML and/or the JSBML
- * library.
- * 
- * @author robert
- * @see org.insilico.testplug
- * @see org.sbml.jsbml
- *
- */
-package org.insilico.testplug.services;
