--- bundles/org.insilico.testplug/src/org/insilico/testplug/services/provider/SBMLEditorProvider.java
@@ -1,26 +1,0 @@
-package org.insilico.testplug.services.provider;
-
-import org.eclipse.fx.code.editor.fx.e4.EditorClassURLProvider;
-import org.insilico.testplug.editor.DocumentTreeView;
-import org.osgi.service.component.annotations.Component;
-
-/**
- * OSGi Service Component which provides a editor part for SBML files. The editor is provides via
- * the {@link EditorClassURLProvider} service.
- * 
- * @author roman
- *
- */
-@Component
-public class SBMLEditorProvider implements EditorClassURLProvider {
-
-    @Override
-    public boolean test(String t) {
-        return t.endsWith("sbml");
-    }
-
-    @Override
-    public String getBundleClassURI(String uri) {
-        return "bundleclass://org.insilico.testplug/" + DocumentTreeView.class.getName();
-    }
-}
