--- bundles/org.insilico.testplug/src/org/insilico/testplug/services/provider/package-info.java
@@ -1,7 +1,0 @@
-/**
- * This package contains OSGi service components which provides other services.
- * 
- * @author robert
- * @see org.insilico.testplug
- */
-package org.insilico.testplug.services.provider;
