--- features/org.insilico.testplug.feature/pom.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
-	xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-	<parent>
-		<groupId>org.insilico</groupId>
-		<artifactId>org.insilico.features</artifactId>
-		<version>1.0.0-SNAPSHOT</version>
-		<relativePath>../</relativePath>
-	</parent>
-	<modelVersion>4.0.0</modelVersion>
-	<artifactId>org.insilico.testplug.feature</artifactId>
-	<packaging>eclipse-feature</packaging>
-	<name>Test Feature</name>
-</project>
