+++ features/pom.xml
@@ -16,8 +16,8 @@
 
 	<modules>
 		<module>org.insilico.core.feature</module>
-		<module>org.insilico.testplug.feature</module>
 		<module>org.insilico.jsbml.feature</module>
-		<module>org.insilico.ui.feature</module>
+		<module>org.insilico.ui.feature</module>		
+		<module>org.insilico.sbmlsheets.feature</module>
 	</modules>
 </project>
